Version 64 bits
git-svn-id: svn://svn.code.sf.net/p/moka-modeller/code@14 578f8393-57f1-403d-81f3-a3b3dc990742

+++ src/lib-controler-gmap/precompile/precompile-face.cc
@@ -200,13 +200,13 @@ void CPrecompileFace::setRandomColor(CDart* ADart, int ADimension)
     }
   else
     { 
-      int c = 0;
+      long int c = 0;
       
       CCoverage* cov = FParameterGMapV->getMap()->
 	getDynamicCoverage(ADart, ORBIT_CELL[ADimension]);
       
       for ( ; cov->cont(); ++ *cov)
-	c += (int) ** cov;
+	c += (long int) ** cov;
       
       delete cov;
